+++ mypy.ini
@@ -38,7 +38,7 @@ files =
     test/test_type_info.py,
     test/test_utils.py,
     tools/linter/clang_format_utils.py,
-    tools/linter/clang_tidy.py,
+    tools/linter/clang_tidy,
     tools/generate_torch_version.py,
     tools/render_junit.py,
     tools/stats
