+++ tools/git-pre-commit
@@ -7,7 +7,7 @@ python tools/linter/flake8_hook.py
 if [ $(which clang-tidy) ]
 then
   echo "Running pre-commit clang-tidy"
-  python tools/linter/clang_tidy.py \
+  python tools/linter/clang_tidy \
     --paths torch/csrc \
     --diff HEAD \
     -g"-torch/csrc/jit/passes/onnx/helper.cpp" \
