+++ tools/linter/clang_tidy/requirements.txt
@@ -1,0 +1,1 @@
+unidiff==0.6.0
