+++ tools/linter/clang_tidy/run.py
@@ -14,7 +14,6 @@ glob or regular expressions.
 
 
 
-import argparse
 import collections
 import fnmatch
 import json
@@ -267,93 +266,7 @@ def apply_nolint(fname: str, warnings: Dict[int, Set[str]]) -> None:
         f.write("".join(lines))
 
 
-def parse_options() -> Any:
-    """Parses the command line options."""
-    parser = argparse.ArgumentParser(description="Run Clang-Tidy (on your Git changes)")
-    parser.add_argument(
-        "-e",
-        "--clang-tidy-exe",
-        default="clang-tidy",
-        help="Path to clang-tidy executable",
-    )
-    parser.add_argument(
-        "-g",
-        "--glob",
-        action="append",
-        default=[],
-        help="Only lint files that match these glob patterns "
-        "(see documentation for `fnmatch` for supported syntax)."
-        "If a pattern starts with a - the search is negated for that pattern.",
-    )
-    parser.add_argument(
-        "-x",
-        "--regex",
-        action="append",
-        default=[],
-        help="Only lint files that match these regular expressions (from the start of the filename). "
-        "If a pattern starts with a - the search is negated for that pattern.",
-    )
-    parser.add_argument(
-        "-c",
-        "--compile-commands-dir",
-        default="build",
-        help="Path to the folder containing compile_commands.json",
-    )
-    parser.add_argument(
-        "--diff-file", help="File containing diff to use for determining files to lint and line filters"
-    )
-    parser.add_argument(
-        "-p",
-        "--paths",
-        nargs="+",
-        default=["."],
-        help="Lint only the given paths (recursively)",
-    )
-    parser.add_argument(
-        "-n",
-        "--dry-run",
-        action="store_true",
-        help="Only show the command to be executed, without running it",
-    )
-    parser.add_argument("-v", "--verbose", action="store_true", help="Verbose output")
-    parser.add_argument(
-        "--config-file",
-        help="Path to a clang-tidy config file. Defaults to '.clang-tidy'.",
-    )
-    parser.add_argument(
-        "-k",
-        "--keep-going",
-        action="store_true",
-        help="Don't error on compiler errors (clang-diagnostic-error)",
-    )
-    parser.add_argument(
-        "-j",
-        "--parallel",
-        action="store_true",
-        help="Run clang tidy in parallel per-file (requires ninja to be installed).",
-    )
-    parser.add_argument(
-        "--print-include-paths",
-        action="store_true",
-        help="Print the search paths used for include directives"
-    )
-    parser.add_argument(
-        "-I",
-        "--include-dir",
-        action="append",
-        help="Add the specified directory to the search path for include files",
-    )
-    parser.add_argument("-s", "--suppress-diagnostics", action="store_true",
-                        help="Add NOLINT to suppress clang-tidy violations")
-    parser.add_argument(
-        "extra_args", nargs="*", help="Extra arguments to forward to clang-tidy"
-    )
-    return parser.parse_args()
-
-
-def main() -> None:
-    options = parse_options()
-
+def run(options: Any) -> None:
     # This flag is pervasive enough to set it globally. It makes the code
     # cleaner compared to threading it through every single function.
     global VERBOSE
@@ -400,6 +313,3 @@ def main() -> None:
     for line in clang_tidy_output.splitlines():
         if line.startswith(pwd):
             print(line[len(pwd):])
-
-if __name__ == "__main__":
-    main()
